requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
